Script: repo.sh updated for new defaults/ path

+++ repo.sh
@@ -21,12 +21,23 @@ CLOUDBOX_PATH="$HOME/cloudbox"
 CLOUDBOX_REPO="https://github.com/Cloudbox/Cloudbox.git"
 
 ## Clone Cloudbox and pull latest commit
-if [ -d "$CLOUDBOX_PATH/.git" ]; then
-    cd "$CLOUDBOX_PATH"
-    git fetch
-    git reset --hard origin/$(git rev-parse --abbrev-ref HEAD)
-    git submodule init
-    git submodule update
+if [ -d "$CLOUDBOX_PATH" ]; then
+    if [ -d "$CLOUDBOX_PATH/.git" ]; then
+        cd "$CLOUDBOX_PATH"
+        git fetch
+        git reset --hard origin/$(git rev-parse --abbrev-ref HEAD)
+        git submodule init
+        git submodule update
+    else
+        cd "$CLOUDBOX_PATH"
+        git init
+        git remote add origin "$CLOUDBOX_REPO"
+        git fetch
+        git branch master origin/master
+        git checkout -f master
+        git submodule init
+        git submodule update
+    fi
 else
     git clone --recursive "$CLOUDBOX_REPO" "$CLOUDBOX_PATH"
     cd "$CLOUDBOX_PATH"
@@ -36,8 +47,8 @@ fi
 
 ## Copy initial config files
 shopt -s nullglob
-for i in *.default; do
-    if [ ! -f "${i%.*}" ]; then
-        cp -n "${i}" "${i%.*}"
+for i in "$CLOUDBOX_PATH"/defaults/*.default; do
+    if [ ! -f "$CLOUDBOX_PATH/$(basename "${i%.*}")" ]; then
+        cp -n "${i}" "$CLOUDBOX_PATH/$(basename "${i%.*}")"
     fi
 done
